store interfaces separated
